Simplify the steam API.
Change the stream API so that the caller doesn't have to group things by register. Instead the stream module now does that de-dup.
This makes things a bit neater at the call sites, but tbh was a lot of work and probably has some bugs.
I think it will prove useful to have the observers know their reg eventually.