Free up obsservers and streams properly.
This adds a stream_delete() function which will delete a stream and all
its observers. stream_new() was added for neatness.
Observers are now deletable too, in a psudo-virtual way, where the sub-
class can register a function to clean itself up and then the base is
free()d.
Also added FE calling of this new stuff.