Fix buffer handling in stream_frame.c
data_len must be equal to arg_len, so we can just use arg_len instead
and assert that fact for good measure.
This means we don't have to wait for the first frame to know how much to
allocate; it can happen earlier.
The second allocation in that loop probably came from the mistaken assumption
that cb will free it - not the case. How the two ended up there together anyone
knows...